[lib] Moved `_render_image()` into `TermImage`
- Change: Moved render-style-dependent `_render_image()` from `BaseImage` into `TermInage`.
- Change: Moved constants `_LOWER_PIXEL` and `_UPPER_PIXEL` from `.image.common` to `.image.term`.
